+++ spec/support/helpers/links_helper.rb
@@ -10,7 +10,7 @@ module LinksHelper
     last_link = all('a.delete > .fas.fa-circle-xmark.fa-reverse', visible: false).last
     scroll_to_element(last_link)
     # Use Javascript to click link element, as it appears to be size 0
-    execute_script("arguments[0].click();", last_link);
+    execute_script('arguments[0].click();', last_link)
   end
 
   def scroll_to_element(element)